[step3] auth -> member 의존성 삭제
