+++ src/main/java/nextstep/auth/domain/UserDetailsService.java
@@ -2,5 +2,7 @@ package nextstep.auth.domain;
 
 public interface UserDetailsService {
 
+    UserDetails loadByUserEmail(String email, int age);
+
     UserDetails loadByUserEmail(String email);
 }
