+++ src/main/java/nextstep/auth/ui/OauthController.java
@@ -1,13 +1,9 @@
 package nextstep.auth.ui;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.member.application.MemberService;
-import nextstep.auth.application.OauthClientService;
-import nextstep.auth.application.TokenService;
-import nextstep.auth.application.dto.OauthTokenRequest;
-import nextstep.auth.application.dto.ResourceResponse;
+import nextstep.auth.application.OauthFacadeService;
 import nextstep.auth.application.dto.ApplicationTokenResponse;
-import nextstep.member.domain.Member;
+import nextstep.auth.application.dto.OauthTokenRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,15 +15,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class OauthController {
 
-    private final OauthClientService oauthClientService;
-    private final MemberService memberService;
-    private final TokenService tokenService;
+    private final OauthFacadeService oauthFacadeService;
 
     @PostMapping("/github")
     public ResponseEntity<ApplicationTokenResponse> getAccessToken(@RequestBody OauthTokenRequest request) {
-        ResourceResponse resourceResponse = oauthClientService.authenticate(request.getCode());
-        Member member = memberService.findMemberByUserResource(resourceResponse.getEmail(), resourceResponse.getAge());
-        ApplicationTokenResponse applicationTokenResponse = tokenService.createToken(member.getEmail(), member.getPassword());
+        ApplicationTokenResponse applicationTokenResponse = oauthFacadeService.getApplicationToken(request.getCode());
         return ResponseEntity.ok(applicationTokenResponse);
     }
 }
