+++ src/test/java/nextstep/member/ui/AuthenticationPrincipalArgumentResolverTest.java
@@ -1,7 +1,7 @@
 package nextstep.member.ui;
 
 import nextstep.auth.application.JwtTokenProvider;
-import nextstep.auth.domain.LoginMember;
+import nextstep.member.domain.LoginMember;
 import nextstep.auth.domain.UserDetails;
 import nextstep.auth.domain.UserDetailsService;
 import nextstep.auth.exception.AuthenticationException;
